Refactor api.py: extract User-Profile + Relationship routes (#368)
Continues the api.py split. Stacks on top of #367.

- /user-profiles + entries CRUD (5 endpoints) → routes/user_profiles.py
- Per-agent profile visibility (2 endpoints)
- /user-profiles/{chat_id}/relationships (3 endpoints)

Total: 11 endpoints moved.

The `user_profiles: UserProfileStore` instance itself stays in api.py
because `_kb_ingest_people_profiles` (the helper that snapshots people
profiles into the KB) also reads from it. The routes module receives
both `user_profiles` and `agents` via set_dependencies.

api.py: 8,153 → 8,028 lines (-125, -1.5%)
Cumulative since main: 11,138 → 8,028 (-3,110 lines, -27.9%)

Tests: 1542 passed, 1 skipped (federation tests need nacl, pre-existing).
